Changed connect button and implemented message list
Deleted the connect "normal" button and moved it to the dropdown (hope it's correct) menu. Also implemented a listview on the "gallery" fragment with the purpose of showing the messages between the test socket server and the test client (now the Android phone). At least for now only displays whatever the user introduces to the messages box.