Server code improved, implemented error handling and implemented socket client on android app
Server now doesn't destroy socket and associated objects until end of execution or exception and when this loses connection with the client, it restarts itself.

Android app can connect to the server, but it can't yet talk with it. Error handling has also been implemented so the app doesn't just crash.